refactor and added not found and communit feed screens
