+++ src/firebase/clientApp.js
@@ -17,7 +17,7 @@ const firebaseConfig = {
   storageBucket: process.env.REACT_APP_PUBLIC_FIREBASE_STORAGE_BUCKET,
   messagingSenderId: process.env.REACT_APP_PUBLIC_FIREBASE_MESSAGING_SENDER_ID,
   appId: process.env.REACT_APP_PUBLIC_FIREBASE_APP_ID,
-  measurementId: process.env.REACT_APP_PUBLIC_FIREBASE_MESSAGING_SENDER_ID
+  measurementId: process.env.REACT_APP_PUBLIC_FIREBASE_MEASUREMENT_ID
 };
 
 // Initialize Firebase for SSR
