+++ src/navigations/index.js
@@ -11,6 +11,8 @@ import UserHome from '../views/UserScreens/Home/UserHome'
 import Login from '../views/auth/login/Login'
 import SubjectSelection from "../views/UserScreens/SubjectSelection/SubjectSelection";
 import Signup from "../views/auth/SignUp/Signup";
+import CommunityFeed from "../views/UserScreens/CommunityFeed/CommunityFeed";
+import NotFound from "../views/auth/NotFound/NotFound";
 // import LandingLayout from "../layouts/LandingLayout/LandingLayout";
 
 const Navigations = () => {
@@ -22,7 +24,7 @@ const Navigations = () => {
 
                     <Route path="/*" element={
                         <GuestLayout>
-                            <Login />
+                            <NotFound />
                         </GuestLayout>
                     } />
                     <Route path="/login" element={
@@ -39,6 +41,7 @@ const Navigations = () => {
 
                     <Route path="/home" element={<UserLayout><UserHome /></UserLayout>} />
                     <Route path="/subjects" element={<UserLayout><SubjectSelection /></UserLayout>} />
+                    <Route path="/feed" element={<UserLayout><CommunityFeed /></UserLayout>} />
 
                 </RouterRoutes>
             </Suspense>
