+++ src/views/UserScreens/CommunityFeed/CommunityFeed.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const CommunityFeed = () => {
+  return (
+    <div>CommunityFeed</div>
+  )
+}
+
+export default CommunityFeed
