+++ src/views/auth/NotFound/NotFound.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const NotFound = () => {
+  return (
+    <div className='text-9xl'>Page Not Found 😢</div>
+  )
+}
+
+export default NotFound
